kernel: clippy --fix mechanical batch (10 style lints)
Mechanical fixes auto-applied by `cargo clippy --fix` for these lints
(no behaviour change, all build-clean):

  unnecessary_sort_by       3 fixes  .sort_by(|a,b| a.k.cmp(&b.k))
                                     → .sort_by_key(|x| x.k)
  manual_div_ceil          ~16       (a + b - 1) / b → a.div_ceil(b)
  manual_strip                       if s.starts_with(p) { &s[p.len()..] }
                                     → s.strip_prefix(p)
  get_first                          v[0] → v.first()
  manual_contains                    s.iter().any(|x| *x == y) → s.contains(&y)
  partialeq_to_none                  x == None → x.is_none()
  manual_pattern_char_comparison     s.starts_with(|c| c == 'a' || c == 'b')
                                     → s.starts_with(['a','b'])
  unnecessary_cast                   redundant `as Same` removals
  needless_borrow                    extra `&` removals
  redundant_closure                  `|x| f(x)` → `f`
  let_and_return                     `let x = y; x` → `y`
  needless_continue                  unreachable continue removals

51 files touched across kernel/.  Build clean.

+++ kernel/src/fs/fnotify.rs
@@ -197,8 +197,8 @@ pub fn record_event(notify_type: NotifyType, kind: EventKind) -> KernelResult<()
         t.events += 1;
         t.event_counts[event_index(kind)] += 1;
         t.queue_depth += 1;
-        if t.queue_depth > t.max_queue_depth as u32 {
-            t.queue_depth = t.max_queue_depth as u32;
+        if t.queue_depth > t.max_queue_depth {
+            t.queue_depth = t.max_queue_depth;
             t.overflows += 1;
             state.total_overflows += 1;
         }
